app/requirements.txt
black==20.8b1
celery==4.4.7
Django>=3.1.0,<3.2.0
django-celery-beat>=2.2.0,<2.3.0
django-enum-choices>=2.1.3,<2.2.0
djangorestframework>=3.11.0,<3.12.0
djangorestframework-simplejwt>=4.6.0,<4.7.0
flake8>=3.8.3,<3.9.8
isort>=5.6.4,<5.7.0
psycopg2-binary>=2.8.5,<2.9.0
pytest>=6.0.2,<6.1.0
pytest-cov>=2.10.1,<2.11.0
pytest-django>=4.1.0,<4.2.0
pytest-sugar>=0.9.4,<1.0.0
redis>=3.5.3,<3.6.0
requests>=2.25.1,<2.26.0
whitenoise>=5.2.0,<5.3.0
